requirements.txt
gradio==4.31.0
numpy>=2.0.0
trimesh>=4.4.0
huggingface_hub>=0.23.0
python-dotenv>=1.0.1
